Сделана фильтрация полностью
