+++ js/main.js
@@ -5,27 +5,21 @@ import './validation-user-form.js';
 import './user-form.js';
 import './user-form-change-size-image.js';
 import './image-effects.js';
-import { debounce } from './utils.js';
+import { debounce, onFilterClick } from './utils.js';
 
 const RERENDER_DELAY = 500;
 
 const filterContainer = document.querySelector('.img-filters');
 const currentButton = filterContainer.querySelector('.img-filters__button--active');
 filterContainer.classList.remove('img-filters--inactive');
-
-
-const onFilterClick = (cb) => {
-  filterContainer.addEventListener('click', (evt) => {
-    const target = evt.target.closest('.img-filters__button');
-    if (target) {
-      cb(target.getAttribute('id'), target);
-    }
-  });
-};
+const picturesContainer = document.querySelector('.pictures');
 
 getData()
   .then((posts) => {
     paintPosts(posts, 'filter-default', currentButton);
-    onFilterClick(debounce((filter, activeButton) => paintPosts(posts, filter, activeButton), RERENDER_DELAY));
+    onFilterClick(debounce((filter, activeButton) => {
+      const eventFunction = paintPosts(posts, filter, activeButton);
+      picturesContainer.removeEventListener('click', eventFunction);
+    }, RERENDER_DELAY));
   });
