+++ js/utils.js
@@ -1,3 +1,5 @@
+const filterContainer = document.querySelector('.img-filters');
+
 const getRandomInteger = (min, max) => { // функция генерации случайного целого числа из диапазона
   const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
   const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
@@ -36,4 +38,14 @@ const debounce = (callback, timeoutDelay) => {
   };
 };
 
-export{isPressedKeyEscape, removeDomElement, createRandomNumberFromRangeGenerator, debounce};
+const onFilterClick = (cb) => {
+  filterContainer.addEventListener('click', (evt) => {
+    const target = evt.target.closest('.img-filters__button');
+    if (target) {
+      cb(target.getAttribute('id'), target);
+    }
+  });
+};
+
+
+export{isPressedKeyEscape, removeDomElement, createRandomNumberFromRangeGenerator, debounce, onFilterClick};
